+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **BC break**: Increased min. PHP version to 7.4.
 - Increased min. Node.js version to 14.15.
+- Fix: App API ESI response did not contain error message in body.
 
 ## 1.27.1